Move the new widget to a separate file
